Add Helper handle to streamline TrackContainer construction

See merge request atlas/athena!65867

+++ Event/xAOD/xAODTracking/xAODTracking/versions/TrackBackendAuxContainer_v1.h
@@ -21,7 +21,7 @@ namespace xAOD {
 
         std::vector<unsigned int> nMeasurements;
         std::vector<unsigned int> nHoles;
-        std::vector<float> chi2;
+        std::vector<double> chi2;
         std::vector<unsigned int> ndf;
         std::vector<unsigned int> nOutliers;
         std::vector<unsigned int> nSharedHits;
